Revert autoplay property support from Samsung media abstraction
Historically, the Samsung media abstraction always autoplays. Support
for setting the 'autoplay' property to false was added in a
previous commit. However, this meant that applications never
received a 'canplay' event after calling load(), but some rely on
it. Attempts to supply the expected events caused further problems,
so we are reverting to the original behaviour. (TVPJSFRMWK-1784)