Handle Samsung 2010 device returning -1 on failure
The Samsung 2010 device returns -1 to indicate a failed API call, rather than 0 (other devices) or false (as documented). Update checks against JumpForward(), JumpBackward() and Pause() calls to take this into account. Refactored and added to the tests to cover the case where devices return false for failure, even though none of the ones we've tested do.