Don't call Stop() when playback completes on Samsung
Previously we called Stop() on the Samsung API and reset our current
time variable when playback completed. This meant that the screen
went black and we didn't get the final playback position in the
'completed' event.

Since we must call Stop() before calling ResumePlay() - otherwise
the API call fails - do this just prior to ResumePlay() when a
previous piece of media has finished playing.