Ensure that seek sentinel giving up doesn't stop pause sentinel
Added a test to ensure that when the seek sentinel gives up trying, the pause
sentinel is then able to run immediately if applicable.

Adding this test revealed that our seek sentinel always assumed that the seek
was going to be successful and updated the 'last sentinel time' member variable
accordingly. This is wrong and undesirable, so we removed the offending line
as a result of the test.