Modify behaviour of enter buffering sentinel
Sentinel enters buffering if the playback position hasn't changed
on the second activation of the sentinel, NOT after waiting for
playback position to have advanced from its original position by
2s.

This caters for the situation that a device exits buffering, but
hasn't actually buffered enough to play >2s of media. Without this
change, the media API would remain in the PLAYING state.