Ensure NO sentinels are fired when time goes backwards
The original test was passing because it was "reversing" time by more than 15 seconds, so the SEEK sentinel was firing rather than the BUFFERING sentinel. This was still wrong! Improve the test to check that no sentinels fire at all (with clearer reasoning for our choice of times), and make a one-line change to the implementation for the new test to pass.