IPTV-673 Moved isBroadcastSupported Device modifier out of base class.
Previously requiring in basetvsource for access to the STATE property would overwrite the Device prototype often unintentionally.
Moved prototype manipulation to "sub classes" of basetvsource (samsungtvsource and tizentvsource).
Updated the tests to reflect the fact that a config that uses the basetvsource as a modifier will not affect the isBroadcastSupported method.
Fixed inconsistent indentation in basetvsource.js