Optimise unit tests
One of the reasons that the unit tests take so long to run is that we are forcing require to reload everything on every test. This is to ensure that require module side effects (e.g. a device modifier changing a function on the Device prototype) apply when necessary.

This commit excludes files that do not need to be reloaded from this forced reloading, hence speeding up the tests.