Reposition element after animating via translate3d()
Ensures that other elements on the page are affected by this element's new
position as they should be. Removing the CSS transform property also
ensures that the element becomes part of the same layer as the rest of the
page after animating, which avoids idiosyncracies with positioning and
z-indexing of child elements, etc.